pyproject.toml
[project]
name = "fastqqbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "qq-botpy>=1.2.1",
    "python-dotenv>=1.1.0",
]
